fix: avoid crashing when selecting a 4-letter currency (#3064)
* fix: crash on changing currency

On the assets page if the user was to change the currency to SATS, LINK or BITS the app would crash. It turns out that we rely on Intl.NumberFormat for currency, but it only works with ISO 4217 - which requires the currency length to be exactly 3 characters. With anything else the call will throw. For now we decided to ignore any non ISO 4217 currencies.

resolves: #3045

* chore: add compojoom on cla allowlist

* chore: add eslint-plugin-no-only-tests plugin

Will help us catch .only flags on tests.